// src/pages/createJob/changeCorporateAccount/edit/constants/types.ts
export type TFullName = {
  title: string;
  firstName: string;
  lastName: string;
};

export type TFullNameForCorporateInfo = [
  {
    title: string;
    firstName: string;
    lastName: string;
  }
];

export type TSubAddress = {
  addressNo: string;
  building?: string;
  floor?: string;
  mooNo?: string;
  soi?: string;
  road?: string;
  tambon: string;
  amphoe: string;
  province: string;
  postalCode: string;
  country: string;
  //type?: number;
};
// ...
export type TAttorney = {
  corporateCode?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  telephone?: string;
  email?: string;
  personalId?: string;
  types?: number;
}

export type TAuthorizePerson = {
  corporateCode?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  types?: number;
  personalId?: string;
};
export type TContactPerson = {
  corporateCode?: string;
  fullNames: TFullName[];
  position: string;
  division: string;
  telephone: string;
  email: string;
  personalId?: string;
};

export type TDirector = {
  corporateCode?: string;
  fullNames: TFullName[];
  citizenId?: string;
  passportId?: string;
  expiryDate: string;
  nationality: string;
  addresses: TSubAddress[];
  types?: number;
  personalId?: string;
};

export type TJuristicsShareholders = {
  corporateCode?: string;
  juristicName: string;
  registrationNo: string;
  registeredCountry: string;
  sharePercentage: number;
  juristicId?: string;
};